Making changes permanent with UserDefaults
ExpenseItem is now codable and can be encoded/decoded to JSON to be stored in the UserDefaults (still this is now to be done in real apps).
Also, add an initializer to load data from the UserDefaults on Expenses creation aka app start up.
Also, used didSet observer to update UserDefaults when there are changes.